requirements.txt
absl-py==2.3.1
astunparse==1.6.3
beautifulsoup4==4.14.2
blinker==1.9.0
bottle==0.13.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clr_loader==0.2.7.post0
colorama==0.4.6
deepface==0.0.95
et_xmlfile==2.0.0
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.75.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
logging==0.4.9.6
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mtcnn==1.0.0
mysql-connector-python==9.4.0
namex==0.1.0
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pickle-mixin==1.0.2
pillow==11.3.0
pluggy==1.6.0
protobuf==6.32.1
proxy_tools==0.1.0
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pywebview==6.0
requests==2.32.5
retina-face==0.0.17
rich==14.1.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
